feat: added 'add note' dialog and refactored dialogs into a separate folder
