+++ src/lib/components/dialogs/delete-book-dialog.svelte
